Adding Jenkins + Github Pr Comment Support.
This will accept post-build POSTS to /_jenkins url.
If the post contains a `repo`, it will look through open
github PR's to find a matching PR (by sha hash) and comment
on that PR with the Jenkins status, the hash it tested, and
a link to the jenkins build.

In order to build jenkins url's properly, you must set the
`PANKY_JENKINS_URL` environment variable to your jenkins host.